Moving response body streaming (#95)
Moving response body streaming from the client into the meta request so that the streaming logic is in one place. In an upcoming change, the meta request will be scheduling more on its io_event_loop, and this will also help keep all tasks for that event loop under the same type.