Fix off-by-one error in calculating the number of parts
When calculating the total number of parts for a given range `start..end` and `part_size`,
`end+1` needs to be used instead of `end`. The reason is that the code consistently uses
Http Byte ranges, which start at index 0 and end at `ContentLength-1`.

The code ensures that `second_part_start` is aligned at `part_size`, hence the
number of remaining chunks is `ceil((end + 1 - second_part_start) / part_size)`.