Stop wrapping synchronous stream in an async stream, to make how things work more obvious.
The number of prepare_requests() calls to have in flight at a time is affected by whether we're ACTUALLY doing async streaming. Wrapping synchronous streams in an async stream, then having a separate bool to track whether it was ACTUALLY async was really confusing. So just have 2 separate streams, async and synchronous, and the one function that reads from them can just have an obvious if/else way of doing things.