Fix bug where we could stall due to mis-counting num_parts_read when resuming a canceled upload.
Bug was: When resuming a paused upload, we increment num_parts_started/completed to account for parts we don't need to re-upload. But we forgot to update `num_parts_read`.

Updating the docs on `num_parts_read` to reflect this new meaning was ... pretty confusing "number of parts started, and read from stream, or that we don't need to read from stream, btw doesn't necessarily reflect the index of where we are due to skipping etc arglbargl"

Decided to just rename the variable to be what we actually care about: ~num_parts_read~ -> `num_parts_pending_read`. Now it goes up and down (not just up) and means what it says and is hopefully less bug-prone.