Allow disabling upload checksums while using upload review
Some clients (Mountpoint) might want to disable actually sending upload
checksums when talking to S3 endpoints that don't support the additional
checksums feature. But they still want to use the upload review callback
to at least check that the parts sent matched the checksums they expect.
This change makes it possible to set a checksum location of AWS_SCL_NONE
while still computing checksums for each upload part. The upload review
callback will be invoked as normal, but the checksums won't be included
in the requests sent to S3.

Signed-off-by: James Bornholt <bornholt@amazon.com>